refactor(log): move log types, hook and utils into modules/log with re-exports
